Markers now all display and function properly. The Spinner works as well!
Added Spinner to MapsActivity.java and activity_maps.xml
Added places_array.xml with places on campus
Added Markers to all locations on campus in MapsActivity.java, and creates an ArrayList of the markers. The separator lines ("-----") have null items in the ArrayList.
The Spinner accesses Marker items in the ArrayList to select the Markers. The positions of the null items are excluded, and nothing will be done if they are selected.